+++ src/app/register/aina/page.tsx
@@ -1,6 +1,6 @@
 import { Aina } from "@/lib/types";
 import { db } from "../../../../db/kysely/client";
-import RegistrationStep2 from "../components/RegistrationStep2";
+import RegistrationStep2 from "../../../components/auth/RegistrationStep2";
 
 export default async function AinaRegistration() {
   const ainaList: Aina[] = await db.selectFrom('aina').select(['id', 'name', 'created_at as createdAt']).execute();
