+++ src/components/auth/RegistrationStep1.tsx
