+++ src/components/auth/RegistrationStep2.tsx
